examples/comprehensive_async_example/requirements.txt
# Requirements for comprehensive async example
python-dotenv>=1.0.0
asyncio-timeout>=4.0.0  # For timeout handling in older Python versions
